Simulation creation and execution
Server now has API to create and run simulations. Client is able to
run that simulation.

For now the simulation configuration in the client is fixed, and after
creating the simulation the simulation is immediately executed.

Added server test to ensure operation sequence runs correctly. And
updated other tests and implementations to store simulation by ID.